[refactor] add parser package for parser classes
